Added database support. Selenium needs lots more improvements to speed and stability (check if transcript button even exists, better clicking decisions).
